+++ app/workers/consolidate_study_configuration_for_study_worker.rb
@@ -1,0 +1,15 @@
+class ConsolidateStudyConfigurationForStudyWorker
+  include Sidekiq::Worker
+
+  sidekiq_options(queue: :default, retry: 5)
+
+  def perform(study_id, version = nil, user_id = nil)
+    ::PaperTrail.whodunnit = user_id
+    Study::ConsolidateStudyConfiguration.call(
+      study_id: study_id,
+      version: version
+    )
+  ensure
+    ::PaperTrail.whodunnit = nil
+  end
+end