+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### Fixed
 
+* <[ERICA-10](https://jira.pharmtrace.com/browse/ERICA-10)> Fix
+  consolidation of visit types, required series and technical qc
+  results after locking the study configuration at a specific version
 * Fix passing down correct version parameter to retrieve the correct
   study configuration (affects Upload of Study Configuration)